CygnetCI.Api/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.49
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
configparser==7.2.0
python-multipart==0.0.6
